refactored date cleaning
